Fix RDS instance resource
