+++ docs/resources/rds_instance.md
@@ -127,7 +127,6 @@ The following arguments are supported:
 * `name` - (Required, String) Specifies the DB instance name. The DB instance name of the same type
   must be unique for the same tenant. The value must be 4 to 64 characters in length and start with a letter.
   It is case-sensitive and can contain only letters, digits, hyphens (-), and underscores (_).
-  Changing this parameter will create a new resource.
 
 * `flavor` - (Required, String) Specifies the specification code.
